# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Upcreate Creator Intelligence API",
    "description": "Backend API for creator metrics, commerce scoring, brands, and campaigns (MVP).",
    "version": "0.1.0"
  },
  "servers": [
    { "url": "/api/v1", "description": "Relative (same host)" },
    { "url": "http://localhost:4000/api/v1", "description": "Local dev" }
  ],
...
  "paths": {
...
    "/creators": {
      "get": {
        "tags": ["Creators"],
        "summary": "List creators",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 100 } },
          {
            "name": "platform",
            "in": "query",
            "schema": { "type": "string", "enum": ["INSTAGRAM", "TIKTOK"] }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Use commerceScore to sort by commerce score descending",
            "schema": { "type": "string", "enum": ["commerceScore"] }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated creators with nested social accounts (each with metrics) and commerce score",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreatorListResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Creators"],
        "summary": "Create creator",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCreator" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Creator" }
              }
            }
          },
          "400": { "description": "Validation error" },
          "409": { "description": "Creator with this email already exists" }
        }
      }
    },
    "/creators/{id}": {
      "get": {
        "tags": ["Creators"],
        "summary": "Get creator by ID",
        "parameters": [{ "$ref": "#/components/parameters/CreatorId" }],
        "responses": {
          "200": {
            "description": "Creator with nested accounts; each account includes metrics and recent posts",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreatorDetail" }
              }
            }
          },
          "404": { "description": "Not found" }
        }
      },
      "put": {
        "tags": ["Creators"],
        "summary": "Update creator",
        "parameters": [{ "$ref": "#/components/parameters/CreatorId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCreator" }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Creator" }
              }
            }
          },
          "404": { "description": "Not found" }
        }
      },
      "delete": {
        "tags": ["Creators"],
        "summary": "Delete creator",
        "parameters": [{ "$ref": "#/components/parameters/CreatorId" }],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "message": { "type": "string" } }
                }
              }
            }
          },
          "404": { "description": "Not found" }
        }
      }
    },
    "/creators/{id}/accounts": {
      "post": {
        "tags": ["Creators"],
        "summary": "Add social account",
        "parameters": [{ "$ref": "#/components/parameters/CreatorId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddSocialAccount" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SocialAccount" }
              }
            }
          },
          "400": { "description": "Validation error" },
          "404": { "description": "Creator not found" },
          "409": { "description": "Account with this platform/username already exists" }
        }
      }
    },
    "/creators/{id}/accounts/{accountId}": {
      "delete": {
        "tags": ["Creators"],
        "summary": "Remove social account",
        "parameters": [
          { "$ref": "#/components/parameters/CreatorId" },
          { "$ref": "#/components/parameters/SocialAccountId" }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "message": { "type": "string" } }
                }
              }
            }
          },
          "404": { "description": "Social account not found" }
        }
      }
    },
    "/creators/{id}/accounts/{accountId}/ingest": {
      "post": {
        "tags": ["Creators"],
        "summary": "Trigger ingestion for a specific social account",
        "parameters": [
          { "$ref": "#/components/parameters/CreatorId" },
          { "$ref": "#/components/parameters/SocialAccountId" }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "message": { "type": "string" } }
                }
              }
            }
          },
          "404": { "description": "Social account not found" }
        }
      }
    },
    "/creators/{id}/calculate-score": {
      "post": {
        "tags": ["Creators"],
        "summary": "Recalculate commerce score",
        "parameters": [{ "$ref": "#/components/parameters/CreatorId" }],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CommerceScore" }
              }
            }
          },
          "404": { "description": "Not found" }
        }
      }
    },
...
    "/matching/campaigns/{campaignId}/suggestions": {
      "get": {
        "tags": ["Matching"],
        "summary": "Ranked creator suggestions for campaign",
        "parameters": [
          { "name": "campaignId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "platform", "in": "query", "schema": { "$ref": "#/components/schemas/Platform" } },
          { "name": "minCommerceScore", "in": "query", "schema": { "type": "number" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Ranked list with fitScore" } }
      }
    },
    "/matching/creators/{creatorId}/intelligence/refresh": {
      "post": {
        "tags": ["Matching"],
        "summary": "Recompute scale, lastPostAt, commerce score",
        "parameters": [{ "name": "creatorId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Creator with metrics" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Login via POST /auth/login; send Authorization: Bearer <token>"
      }
    },
    "parameters": {
      "CreatorId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
      "SocialAccountId": {
        "name": "accountId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
...
    },
    "schemas": {
...
      "Platform": { "type": "string", "enum": ["INSTAGRAM", "TIKTOK"] },
...
      "Creator": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email", "nullable": true },
          "avatarUrl": { "type": "string", "nullable": true },
          "bio": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "SocialAccount": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "creatorId": { "type": "string" },
          "platform": { "$ref": "#/components/schemas/Platform" },
          "username": { "type": "string" },
          "profileUrl": { "type": "string", "nullable": true },
          "followerCount": { "type": "integer" },
          "isConnected": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "SocialAccountWithMetrics": {
        "allOf": [
          { "$ref": "#/components/schemas/SocialAccount" },
          {
            "type": "object",
            "properties": {
              "metrics": {
                "nullable": true,
                "allOf": [{ "$ref": "#/components/schemas/AccountMetrics" }]
              }
            }
          }
        ]
      },
      "SocialAccountDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/SocialAccountWithMetrics" },
          {
            "type": "object",
            "properties": {
              "posts": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/ContentPost" }
              }
            }
          }
        ]
      },
      "ContentPost": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "accountId": { "type": "string" },
          "postId": { "type": "string", "nullable": true },
          "caption": { "type": "string", "nullable": true },
          "mediaType": { "type": "string", "nullable": true },
          "likes": { "type": "integer" },
          "commentsCount": { "type": "integer" },
          "views": { "type": "integer" },
          "postedAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "AccountMetrics": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "accountId": { "type": "string" },
          "avgLikes30d": { "type": "number" },
          "avgComments30d": { "type": "number" },
          "avgViews30d": { "type": "number" },
          "engagementRate": { "type": "number" },
          "contentTopics": { "type": "array", "items": { "type": "string" } },
          "calculatedAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "CommerceScore": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "creatorId": { "type": "string" },
          "engagementScore": { "type": "number" },
          "intentScore": { "type": "number" },
          "commerceScore": { "type": "number" },
          "calculatedAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "CreatorListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CreatorListItem" }
          },
          "meta": { "$ref": "#/components/schemas/PaginationMeta" }
        }
      },
      "CreatorListItem": {
        "allOf": [
          { "$ref": "#/components/schemas/Creator" },
          {
            "type": "object",
            "properties": {
              "accounts": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/SocialAccountWithMetrics" }
              },
              "commerceScore": {
                "nullable": true,
                "allOf": [{ "$ref": "#/components/schemas/CommerceScore" }]
              }
            }
          }
        ]
      },
      "CreatorDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/Creator" },
          {
            "type": "object",
            "properties": {
              "accounts": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/SocialAccountDetail" }
              },
              "commerceScore": {
                "nullable": true,
                "allOf": [{ "$ref": "#/components/schemas/CommerceScore" }]
              }
            }
          }
        ]
      },
      "CreateCreator": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "maxLength": 200 },
          "email": { "type": "string", "format": "email" },
          "avatarUrl": { "type": "string", "format": "uri" },
          "bio": { "type": "string" }
        }
      },
      "UpdateCreator": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "maxLength": 200 },
          "email": { "type": "string", "format": "email" },
          "avatarUrl": { "type": "string", "format": "uri" },
          "bio": { "type": "string" }
        }
      },
      "AddSocialAccount": {
        "type": "object",
        "required": ["platform", "username"],
        "properties": {
          "platform": { "$ref": "#/components/schemas/Platform" },
          "username": { "type": "string", "maxLength": 100 },
          "profileUrl": { "type": "string", "format": "uri" },
          "followerCount": { "type": "integer", "minimum": 0 },
          "isConnected": { "type": "boolean" }
        }
      },
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" },
          "totalPages": { "type": "integer" }
        }
      },
...
    }
  }
}